Code cleanup

+++ code/src/spdb/QueryParameterList.cpp
@@ -78,7 +78,7 @@ size_t QueryParameterList::size() const
 
 void QueryParameterList::remove(size_t i)
 {
-    auto itor = m_items.begin() + i;
+    auto itor = m_items.begin() + (int) i;
     SQueryParameter item = *itor;
     m_index.erase(item->name());
     m_items.erase(itor);
